requirements.txt
rshell
esptool
micropython-esp32-stubs
